fix: migrate to ts
