--- src/theme/Drawer/index.js
@@ -1,18 +1,0 @@
-import React from "react";
-import "./style.css";
-
-export default function Drawer(props) {
-  const { SVG } = props;
-
-  return (
-    <div className="svg-box" style={{}}>
-      <svg
-        className="svg-container"
-        style={{ marginTop: "20px", width: SVG.attrs.width, height: SVG.attrs.height }}
-        ref={(element) => {
-          element && element.appendChild(SVG.toSVG());
-        }}
-      ></svg>
-    </div>
-  );
-}
